requirements.txt
Telethon==1.36.0
Pillow==10.4.0
gTTS==2.5.1
deep-translator==1.11.4
motor==3.6.0
thefuzz==0.22.1
python-dotenv==1.0.1
